test(server): consolidate RegisteredFunctions tests, split out Impl tests
Merge the buildForGroup runtime test back into RegisteredFunctions.test.ts
alongside the existing type tests, and move Impl.buildForGroup to its own
Impl.test.ts so the file's subject matches the module under test.
